requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
boto3==1.18.10
botocore==1.21.10
celery==5.1.2
celery-progress==0.1.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-celery-results==2.2.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-import-export==2.5.0
django-storages==1.11.1
djangorestframework==3.12.4
dnspython==2.1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
jmespath==0.10.0
kombu==5.1.0
MarkupPy==1.14
numpy==1.21.1
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.1
Pillow==9.2.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
XlsxWriter==1.4.4
xlwt==1.3.0
